requirements.txt
easydict
numpy
Pillow
PyYAML
scikit_learn
